REFACTORING: distribute files to directories (#125)
Closes #124
