+++ main.cpp
@@ -1,4 +1,4 @@
-#include "source/logger.hpp"
+#include "logger/logger.hpp"
 
 int main() {
     Logger::setupLogging();
