+++ source/device/switch.cpp
@@ -1,9 +1,9 @@
 #include "switch.hpp"
-#include "logger.hpp"
 
 #include <memory>
 
 #include "link.hpp"
+#include "logger/logger.hpp"
 
 namespace sim {
