+++ source/link.cpp
@@ -1,5 +1,5 @@
 #include "link.hpp"
-#include "logger.hpp"
+#include "logger/logger.hpp"
 #include "event.hpp"
 #include "scheduler.hpp"
