+++ test/device/add_link_test.cpp
@@ -3,7 +3,7 @@
 #include <memory>
 #include <random>
 
-#include "device.hpp"
+#include "device/device.hpp"
 #include "link.hpp"
 #include "utils.hpp"
 #include "utils/loop_iterator.hpp"
