+++ test/simulator/utils.cpp
@@ -1,6 +1,6 @@
 #include "utils.hpp"
 
-#include "logger.hpp"
+#include "logger/logger.hpp"
 #include "simulator.hpp"
 
 namespace test {
