+++ test/simulator/utils.hpp
@@ -4,7 +4,7 @@
 #include <memory>
 #include <utility>
 
-#include "device.hpp"
+#include "device/device.hpp"
 #include "simulator.hpp"
 
 namespace test {
