+++ test/switch/flow_mock.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "flow.hpp"
-#include "sender.hpp"
+#include "device/sender.hpp"
 
 namespace test {
