+++ test/switch/link_mock.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "device.hpp"
+#include "device/device.hpp"
 #include "link.hpp"
 
 class LinkMock : public sim::ILink {
