+++ test/switch/test_switch.cpp
@@ -2,13 +2,13 @@
 
 #include <vector>
 
-#include "device.hpp"
+#include "device/device.hpp"
 #include "flow_mock.hpp"
 #include "link.hpp"
 #include "link_mock.hpp"
 #include "packet.hpp"
 #include "receiver_mock.hpp"
-#include "switch.hpp"
+#include "device/switch.hpp"
 
 namespace test {
